updated
all you need to do is put this file in a folder where you want to create images folders and put all your images which you want to assign in a seperate folder and give that folder's path in the script. 
(You might need to change .replace() as per your preference)
all done now run it.